requirements.txt
flask
python-dotenv
pydantic
tensorflow
scikit-learn